// src/types/truncatedText.ts
export interface ITruncate {
    model: ITruncateModel
  }
  
  export interface ITruncateModel {
    id: number
    naim: string
    text: string
    tobottom: any
    naim2: string
    text2: string
    meta_title: string
    meta_description: string
    keywords: string
    url: string
    cat_id: number
    sort: number
    menu: number
  }
  